backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
python-dotenv==1.1.1

SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9

PyJWT==2.10.1
bcrypt==4.1.3
email-validator==2.2.0
python-multipart==0.0.9

slowapi==0.1.9
Pillow==10.4.0
numpy==2.1.1
gTTS==2.5.3
requests==2.32.3

pytest==8.3.3
